Layout widgets: Container example app
