+++ examples/layout/container/lib/main.dart
@@ -1,0 +1,56 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart' show debugPaintSizeEnabled;
+
+void main() {
+  debugPaintSizeEnabled = false; // Set to true for visual layout
+  runApp(MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter layout demo',
+      home: Scaffold(
+        appBar: AppBar(
+          title: Text('Flutter layout demo'),
+        ),
+        body: Center(child: _buildImageColumn()),
+      ),
+    );
+  }
+
+  // #docregion column
+  Widget _buildImageColumn() => Container(
+        decoration: BoxDecoration(
+          color: Colors.black26,
+        ),
+        child: Column(
+          children: [
+            _buildImageRow(1),
+            _buildImageRow(3),
+          ],
+        ),
+      );
+  // #enddocregion column
+
+  // #docregion row
+  Widget _buildDecoratedImage(int imageIndex) => Expanded(
+        child: Container(
+          decoration: BoxDecoration(
+            border: Border.all(width: 10, color: Colors.black38),
+            borderRadius: const BorderRadius.all(const Radius.circular(8)),
+          ),
+          margin: const EdgeInsets.all(4),
+          child: Image.asset('images/pic$imageIndex.jpg'),
+        ),
+      );
+
+  Widget _buildImageRow(int imageIndex) => Row(
+        children: [
+          _buildDecoratedImage(imageIndex),
+          _buildDecoratedImage(imageIndex + 1),
+        ],
+      );
+  // #enddocregion row
+}
