--- src/_includes/code/layout/container/pubspec.yaml
@@ -1,14 +1,0 @@
-name: container
-description: Demonstrates the Container widget.
-
-dependencies:
-  flutter:
-    sdk: flutter
-
-flutter:
-  uses-material-design: true
-  assets:
-    - images/pic1.jpg
-    - images/pic2.jpg
-    - images/pic3.jpg
-    - images/pic4.jpg
